refactor: rename quadruped.py -> Robotdog.py
