+++ test/test_angle.py
@@ -2,7 +2,7 @@ import time
 
 from model.LegController import LegController
 from model.hardware.Motor import Motor
-from model.types.leg import LegPart
+from model.types.types import LegPart
 
 TESTED_LEG_SHOULDER = Motor.FL_SHOULDER
 TESTED_LEG_ELBOW = Motor.FL_ELBOW
